Move vehicle flags to their own header
